update Thread stream to post the first post
uses the method in mysqlQuerier for getting an smf_message with the
same ID_MSG as the smf_topic's ID_FIRST_MSG

changed config to use 100 simultaneous pool connections
if the threshold is set too low, it can create a problem with
asynchronous bottlenecking.  calls to mysql might depend on eachother
and may become locked in waiting if there are not enough available
connections.